Move Numba scan tests to test_scan
